fix: 결과 이미지 저장, 불러오기 및 차트 오류 수정
